requirements.txt
asgiref==3.7.2
Django==4.1.5
gunicorn==20.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
mysqlclient>=2.0
skyfield
djangorestframework==3.14.0